pdf_AI_Bot/requirements.txt
python-dotenv
huggingface_hub
pypdf2
faiss-cpu
openai
streamlit
langchain
InstructorEmbedding 
sentence_transformers